#99 refactor: Home ProfileList 영역 모듈화
